// apps/frontend/src/api/generated/model/burnUpStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.32
 */

export type BurnUpStatus = (typeof BurnUpStatus)[keyof typeof BurnUpStatus]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BurnUpStatus = {
  ahead: 'ahead',
  on_track: 'on_track',
  behind: 'behind',
  at_risk: 'at_risk',
  off_track: 'off_track',
} as const