// apps/frontend/src/api/generated/model/budgetOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.32
 */
import type { BudgetOutOwnerId } from './budgetOutOwnerId'
import type { BudgetPeriod } from './budgetPeriod'
import type { BudgetOutEndDate } from './budgetOutEndDate'
import type { BudgetMethod } from './budgetMethod'
import type { ExpectedIncomeStrategy } from './expectedIncomeStrategy'
import type { BudgetOutExpectedIncome } from './budgetOutExpectedIncome'
import type { BudgetOutPayPeriodIncomeSourceId } from './budgetOutPayPeriodIncomeSourceId'
import type { BudgetOutEffectiveTo } from './budgetOutEffectiveTo'

export interface BudgetOut {
  id: string
  budget_group_id: string
  household_id: string
  owner_id: BudgetOutOwnerId
  name: string
  period: BudgetPeriod
  start_date: string
  end_date: BudgetOutEndDate
  method: BudgetMethod
  expected_income_strategy: ExpectedIncomeStrategy
  expected_income: BudgetOutExpectedIncome
  currency: string
  income_rolling_periods: number
  scope_accounts: unknown[]
  scope_categories: unknown[]
  scope_tags: unknown[]
  pay_period_income_source_id: BudgetOutPayPeriodIncomeSourceId
  effective_from: string
  effective_to: BudgetOutEffectiveTo
  created_at: string
  updated_at: string
}