// apps/frontend/src/api/generated/model/expectedEventOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.34
 */
import type { Cadence } from './cadence'
import type { ExpectedEventOutMerchantName } from './expectedEventOutMerchantName'
import type { ExpectedEventOutExceptionType } from './expectedEventOutExceptionType'
import type { ExpectedEventOutOverrideAmount } from './expectedEventOutOverrideAmount'
import type { ExpectedEventOutOverrideDate } from './expectedEventOutOverrideDate'

export interface ExpectedEventOut {
  recurrence_id: string
  account_id: string
  expected_date: string
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  expected_amount: string
  currency: string
  cadence: Cadence
  merchant_name: ExpectedEventOutMerchantName
  exception_type?: ExpectedEventOutExceptionType
  override_amount?: ExpectedEventOutOverrideAmount
  override_date?: ExpectedEventOutOverrideDate
}